Split out startEditingExistingPost
